opal-ci: Add Ubuntu 18.04
Start doing builds on latest LTS. We can simplify the build script a
little, as we know that this version of Ubuntu ships with an ARM cross
compiler and a Qemu that can run the tests.

This provides us with:

 gcc 7.3.0 (for native and cross-builds)
 qemu 2.11
 clang 6

Simplify the docker script a little by merging the two different install
lines, and getting rid of the duplicate 'apt-get update' steps. We now
install clang in preparation for future shenanigans, and use the system
qemu instead of building one ourselves.

Signed-off-by: Joel Stanley <joel@jms.id.au>
[stewart: allow 18.04 to fail as we stabilise it in travis]
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>