hmi: Fix clearing HMER on debug trigger
In the recent patch:
  eddff9bf40 hmi: Clear unknown debug trigger

I rebased the code from an older skiboot before the HMI rework. When I
did this, I missed the handled flag. Without this the HMER is not
cleared properly and the HMI keeps happening.

This properly sets the handled flag and hence clears the HMER bit.

Signed-off-by: Michael Neuling <mikey@neuling.org>
Reviewed-by: Mahesh Salgaonkar <mahesh@linux.vnet.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>