AMI BMC: use 0x3a as OEM command
[ Upstream commit e9227752364564581d591d0ecb5b56e7e9bc4244 ]

The 0x3a OEM command is for IBM commands, while 0x32 was for AMI ones.
Sometime in the P8 timeframe, AMI BMCs were changed to listen for our
commands on either 0x32 or 0x3a. Since 0x3a is the direction forward,
we'll use that, as P9 machines with AMI BMCs probably also want these
to work, and let's not bet that 0x32 will continue to be okay.

Suggested-by: Joseph Reynolds <jrey@us.ibm.com>
Suggested-by: Maury Zipse <zipse@us.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>