core: Add test for PCI quirks
Ensure that quirks are run (or not) for given PCI vendor and device IDs.
This tests the quirk infrastructure and the PCI_VENDOR_ID() and
PCI_DEVICE_ID() macros, the latter of which was recently found to be
broken.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>
(cherry picked from commit dc24a1fd61e0b3fbceb1027b2c458bde0257fb38)
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>