init, occ: Initialise OCC earlier on BMC systems
We need to use the OCC to obtain presence data for the SXM2 slots on
Witherspoon systems. This is needed to determine device type for NVLink
GPUs and OpenCAPI devices which can be plugged into the same slot. Support
for this will be implemented in a future patch.

Currently, OCC initialisation is done just before handing over to Linux,
which is well after NPU probe. On FSP systems, OCC boot starts very late,
so we wait until the last possible moment to initialise the skiboot side in
order to give it the maximum time to boot. On BMC systems, OCC boot starts
earlier, so there aren't any issues in moving it earlier in the skiboot
init sequence.

When running on a BMC machine, call occ_pstates_init() as early as
possible in the init sequence. On FSP machines, continue to call it from
its current location.

Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Reviewed-by: Shilpasri G Bhat <shilpa.bhat@linux.vnet.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>