core/pci: Print 'PCI Summary' at PR_NOTICE
The actual entries of the PCI Summary are printed at PR_NOTICE so that
they go to the console during boot. The header however does not which
breaks my patented "grep 'PCI Summary' -A 100" technique for scraping
the summary out of a log file when that log is recorded from the SOL
console.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>