hw/npu2-hw-procedures: Enable RX auto recal on OpenCAPI links
The RX_RC_ENABLE_AUTO_RECAL flag is required on OpenCAPI but not NVLink.

Traditionally, Hostboot sets this value according to the machine type.
However, now that Witherspoon supports both NVLink and OpenCAPI, it can't
tell whether or not a link is OpenCAPI.

So instead, set it in skiboot, where it will only be triggered after we've
done device detection and found an OpenCAPI device.

Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Reviewed-by: Frederic Barrat <fbarrat@linux.ibm.com>
Acked-by: Reza Arbab <arbab@linux.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>