SBE-p8: Do all sbe timer update with xscom lock held
Without this, on some P8 platforms, we could (falsely) think the SBE timer
had stalled getting the dreaded "timer stuck" message.

The code was doing the mftb() to set the start of the timeout period while
*not* holding the lock, so the 1ms timeout started sometime when somebody
else had the xscom lock.

The simple solution is to just do the whole routine holding the xscom lock,
so do it that way.

Signed-off-by: Stewart Smith <stewart@linux.ibm.com>