nvram: Fix a possible NULL pointer de-ref in nvram_query_eq()
A fault will occur if 'value == NULL' is passed to nvram_query_eq() to
check if a given key doesn't exists in nvram partition. This is an
invalid use of the API as its only supposed to be used for keys that
exist in nvram and 'value == NULL' is never possible.

Hence this patch adds an assert to the function to flag such a use and
also prevent NULL being passed as an argument to strcmp().

Signed-off-by: Vaibhav Jain <vaibhav@linux.ibm.com>
Suggested-by: Oliver O'Halloran <oohall@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>